Remove GCM support and migrate to FCM
According to https://developers.google.com/cloud-messaging/android/android-migrate-fcm "On April 10, 2018, Google deprecated GCM. The GCM server and client APIs were removed on May 29, 2019."
Also it states that "GCM and FCM SDKs cannot co-exist within an application.", but this library seems to mix them both. For example RNPushNotificationRegistrationService uses GCM to retrieve the push notification token, even if you use Firebase in your app. This seems to cause some undesired effects. For example when using `com.google.android.gms:play-services-gcm:17.0.0` and `com.google.firebase:firebase-messaging:20.1.0`, and using FirebaseInstanceId to retrieve the token somewhere in your app (or from a different library that uses Firebase), it will return a different token from the one returned by the deprected `InstanceID` used in this library. Also the usage of `InstanceID` seems to invalidate the token obtained with `FirebaseInstanceId`, so that if you try to send a notification using that token, Firebase API returns a `NotRegistered` error.